Remove Multiobjective function (#1033)
